% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_chm_file.R
\name{check_chm_file}
\alias{check_chm_file}
\title{Check attributes of .tif file}
\usage{
check_chm_file(handle)
}
\arguments{
\item{handle}{object of class \linkS4class{GDALReadOnlyDataset}, created by
call to \code{\link[rgdal]{GDAL.open}}.}
}
\description{
Check attributes of .tif file
}
