% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_local_maximas.R
\name{local_maxima_catalog}
\alias{local_maxima_catalog}
\title{find_local_maximas wrapper}
\usage{
local_maxima_catalog(files, search.radius)
}
\arguments{
\item{files}{a character vector of paths to tif files containing the canopy
height map.}

\item{search.radius}{the distance around each raster cell that defines the
search neighbourhood.}
}
\value{
A matrix containing the (x,y,z) coordinates of each maximum.
}
\description{
\code{local_maxima_search} is a wrapper around the \code{find_local_maximas}
function, which implements the local maxima finding algorithm in C++.
}
