% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_maximas.R
\name{local_maxima_search}
\alias{local_maxima_search}
\title{Find locally maximal points in a canopy height map}
\usage{
local_maxima_search(files, search.radius, chm.res = 0.25)
}
\arguments{
\item{files}{a character vector of file paths.}

\item{search.radius}{the radius around each CHM point that will be searched.}

\item{chm.res}{the length of the side of each CHM cell (in metres). Only used
for \code{.laz} files}
}
\value{
a \code{\link[data.table]{data.table}} containing the 3D coordinates
  of the local maxima in the CHM.
}
\description{
\code{local_maxima_search} performs a fixed-radius nearest neighbour search
around each canopy height map (CHM) cell and identifies those cells that are
locally maximal with respect to their height above ground.
}
\details{
The search can be performed on either \code{.laz} (or \code{.las}) files,
containing the LiDAR point-cloud, or on a \code{.tif} file containing the CHM
raster. Note that when a point-cloud is provided, the CHM must be generated
from the files imported and that this will take additional time.
}
