% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_maximas.R
\name{local_maximas_tif}
\alias{local_maximas_tif}
\title{Local maxima search of CHM .tif file}
\usage{
local_maximas_tif(file, search.radius)
}
\arguments{
\item{file}{path to a single .tif file}

\item{search.radius}{the radius around each CHM point that will be searched.}
}
\value{
a \code{\link[data.table]{data.table}} containing the 3D coordinates
  of the local maxima in the CHM.
}
\description{
Local maxima search of CHM .tif file
}
