% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tiff.R
\name{test_read_tiff}
\alias{test_read_tiff}
\title{Test function for reading data from raster file}
\usage{
test_read_tiff(file)
}
\arguments{
\item{file}{path to a .tif file.}
}
\value{
numeric vector containing the raster data
}
\description{
Test function for reading data from raster file
}
